requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.5
Django==4.1.5
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
exceptiongroup==1.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
prompt-toolkit==3.0.36
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
psycopg2-binary==2.8.6
django-environ==0.4.5
